Add CONTRIBUTING.md and supporting local setup files
We want to make sure it's easy for others to assist contributing the
doco, and we also want to make sure people can preview their work before
they submit it. So I've outlined how that is done in CONTRIBUTING.md and
added the supporting files to be able to run a Jekyll build and site
locally.