Provide a more professional site
The previous setup was very bare bones and presented new comers to the
site with a wall of text. It was visually unappealing and difficult to
find content.

To address this, a more feature full theme has been selected and then
additional pages and configuration added to support.

There's more work that can be done, but this is the big first step.